Continue working on Scene System.
Scene system is almost done. Im thinking about functions to update a scene uniform every frame and maybe some other helper functions but for the most part the scene system is done. After this I will probably take some time to make sure its stable and to tinker around to see if I cant push this to go a little faster. Hopefully after that Ill continue implementation of features in the LearnOpenGL tutorial