Implement Materials
Implement part of a crude material system. In the future a single class should handle all of this but for now manually setting uniforms will work. Also implement light controls as uniforms.

Next up would be textures and specular maps but because textures are already here and just need some mild rearranging specular maps should be most of the work.

I also changed the background color.

This project has steered more back into the tutorial space which is fine by me but I want something more flexible and customizable in the future.  I cant claim to be at the end of the tutorial but with what I have now I think I could get a good start on the project next. That said I want to see this project until model loading.